Revert "Fb/ft tailwind pages layout (#1)"
This reverts commit de3ec225143e8999c798695f9d6c31defa1ff16d.
