--- .prettierrc.json
@@ -1,11 +1,0 @@
-{
-    "semi": false,
-    "trailingComma": "es5",
-    "singleQuote": true,
-    "tabWidth": 4,
-    "useTabs": false,
-    "quoteProps": "as-needed",
-    "bracketSpacing": true,
-    "bracketSameLine": true,
-    "printWidth": 90
-}
