+++ eslint.config.mjs
@@ -1,16 +1,16 @@
-import { dirname } from 'path'
-import { fileURLToPath } from 'url'
-import { FlatCompat } from '@eslint/eslintrc'
+import { dirname } from "path";
+import { fileURLToPath } from "url";
+import { FlatCompat } from "@eslint/eslintrc";
 
-const __filename = fileURLToPath(import.meta.url)
-const __dirname = dirname(__filename)
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 const compat = new FlatCompat({
-    baseDirectory: __dirname,
-})
+  baseDirectory: __dirname,
+});
 
 const eslintConfig = [
-    ...compat.extends('next/core-web-vitals', 'next/typescript', 'prettier'),
-]
+  ...compat.extends("next/core-web-vitals", "next/typescript"),
+];
 
-export default eslintConfig
+export default eslintConfig;
