--- src/pages/dashboard/index.tsx
@@ -1,7 +1,0 @@
-import { FunctionComponent } from 'react'
-
-const Dashboard: FunctionComponent = () => {
-    return <div className="text-3xl bg-amber-700 text-primary">this is the dashboard</div>
-}
-
-export default Dashboard
