--- tailwind.config.js
@@ -1,13 +1,0 @@
-export default {
-    theme: {
-        extend: {
-            colors: {
-                primary: '#ff49db',
-            },
-            fontFamily: {
-                sans: ['Helvetica', 'Arial', 'sans-serif'],
-            },
-        },
-    },
-    plugins: [],
-}
